:white_check_mark: removed run_rules!() from all tests.
