+++ src/rules/tests/mail.rs
@@ -14,48 +14,44 @@
  * this program. If not, see https://www.gnu.org/licenses/.
  *
 **/
-#[cfg(test)]
-pub mod test {
-    use crate::{
-        mime::parser::MailMimeParser,
-        rules::{
-            address::Address,
-            rule_engine::{RuleEngine, Status},
-            tests::helpers::get_default_state,
-        },
-        smtp::mail::Body,
-    };
+use crate::{
+    mime::parser::MailMimeParser,
+    rules::{
+        address::Address,
+        rule_engine::{RuleEngine, Status},
+        tests::helpers::get_default_state,
+    },
+    smtp::mail::Body,
+};
 
-    #[test]
-    fn test_mail_from_rules() {
-        crate::receiver::test_helpers::logs::setup_logs();
+#[test]
+fn test_mail_from_rules() {
+    crate::receiver::test_helpers::logs::setup_logs();
 
-        let re =
-            RuleEngine::new("./src/rules/tests/rules/mail").expect("couldn't build rule engine");
+    let re = RuleEngine::new("./src/rules/tests/rules/mail").expect("couldn't build rule engine");
 
-        let mut state = get_default_state();
-        {
-            let email = state.get_context();
-            let mut email = email.write().unwrap();
+    let mut state = get_default_state();
+    {
+        let email = state.get_context();
+        let mut email = email.write().unwrap();
 
-            email.envelop.mail_from = Address::new("staff@viridit.com").unwrap();
-            email.body = Body::Parsed(Box::new(
-                MailMimeParser::default()
-                    .parse(
-                        br#"From: staff <staff@viridit.com>
+        email.envelop.mail_from = Address::new("staff@viridit.com").unwrap();
+        email.body = Body::Parsed(Box::new(
+            MailMimeParser::default()
+                .parse(
+                    br#"From: staff <staff@viridit.com>
 Date: Fri, 21 Nov 1997 10:01:10 -0600
 
 This is a reply to your hello."#,
-                    )
-                    .unwrap(),
-            ));
-        }
-
-        assert_eq!(re.run_when(&mut state, "mail"), Status::Accept);
-        assert_eq!(re.run_when(&mut state, "postq"), Status::Accept);
-        assert_eq!(
-            state.get_context().read().unwrap().envelop.mail_from.full(),
-            "no-reply@viridit.com"
-        );
+                )
+                .unwrap(),
+        ));
     }
+
+    assert_eq!(re.run_when(&mut state, "mail"), Status::Accept);
+    assert_eq!(re.run_when(&mut state, "postq"), Status::Accept);
+    assert_eq!(
+        state.get_context().read().unwrap().envelop.mail_from.full(),
+        "no-reply@viridit.com"
+    );
 }
