+++ src/rules/tests/rules/rcpt/main.vsl
@@ -2,20 +2,18 @@ obj ident "john" "johndoe";
 obj fqdn "viridit" "viridit.com";
 obj addr "customer" "customer@company.com";
 
-run_rules!(
-  #{
-    rcpt: [
-      rule "test_ident" || if john in ctx.rcpt.local_parts { vsl::next() } else { vsl::deny() },
-      rule "test_fqdn" || if viridit in ctx.rcpt.domains { vsl::next() } else { vsl::deny() },
-      rule "test_addr" || if customer in ctx.rcpt { vsl::accept() } else { vsl::deny() },
-    ],
+#{
+  rcpt: [
+    rule "test_ident" || if john in ctx.rcpt.local_parts { vsl::next() } else { vsl::deny() },
+    rule "test_fqdn" || if viridit in ctx.rcpt.domains { vsl::next() } else { vsl::deny() },
+    rule "test_addr" || if customer in ctx.rcpt { vsl::accept() } else { vsl::deny() },
+  ],
 
-    postq: [
-      action "test_rewrite" || {
-        ctx.rewrite_rcpt("johndoe@compagny.com", "johndoe@viridit.com");
-        ctx.remove_rcpt("customer@company.com");
-        ctx.add_rcpt("no-reply@viridit.com");
-      },
-    ]
-  }
-)
+  postq: [
+    action "test_rewrite" || {
+      ctx.rewrite_rcpt("johndoe@compagny.com", "johndoe@viridit.com");
+      ctx.remove_rcpt("customer@company.com");
+      ctx.add_rcpt("no-reply@viridit.com");
+    },
+  ]
+}
