+++ src/rules/tests/types/objects/main.vsl
@@ -1,33 +1,31 @@
 import "vars" as vars;
 
-run_rules!(
-    #{
-        connect: [
-            action "test_print_objects" || {
-                print(vars::ip4);
-                print(vars::ip6);
-                print(vars::rg4);
-                print(vars::rg6);
-                print(vars::address);
-                print(vars::fqdn);
-                print(vars::regex);
-                print(vars::identifier);
-                print(vars::str);
-                print(vars::file);
-                print(vars::group);
+#{
+    connect: [
+        action "test_print_objects" || {
+            print(vars::ip4);
+            print(vars::ip6);
+            print(vars::rg4);
+            print(vars::rg6);
+            print(vars::address);
+            print(vars::fqdn);
+            print(vars::regex);
+            print(vars::identifier);
+            print(vars::str);
+            print(vars::file);
+            print(vars::group);
 
-                print(vars::ip4.to_debug());
-                print(vars::ip6.to_debug());
-                print(vars::rg4.to_debug());
-                print(vars::rg6.to_debug());
-                print(vars::address.to_debug());
-                print(vars::fqdn.to_debug());
-                print(vars::regex.to_debug());
-                print(vars::identifier.to_debug());
-                print(vars::str.to_debug());
-                print(vars::file.to_debug());
-                print(vars::group.to_debug());
-            },
-       ]
-    }
-)
+            print(vars::ip4.to_debug());
+            print(vars::ip6.to_debug());
+            print(vars::rg4.to_debug());
+            print(vars::rg6.to_debug());
+            print(vars::address.to_debug());
+            print(vars::fqdn.to_debug());
+            print(vars::regex.to_debug());
+            print(vars::identifier.to_debug());
+            print(vars::str.to_debug());
+            print(vars::file.to_debug());
+            print(vars::group.to_debug());
+        },
+   ]
+}
