+++ src/rules/tests/types/status/main.vsl
@@ -1,20 +1,18 @@
-run_rules!(
-    #{
-        connect: [
-            rule "test_status" || {
-                if vsl::next() is vsl::next()
-                && vsl::next() == vsl::next()
+#{
+    connect: [
+        rule "test_status" || {
+            if vsl::next() is vsl::next()
+            && vsl::next() == vsl::next()
 
-                && vsl::next() not vsl::accept()
-                && vsl::next() != vsl::accept()
+            && vsl::next() not vsl::accept()
+            && vsl::next() != vsl::accept()
 
-                && vsl::next().to_string() == "next"
-                && vsl::next().to_debug() == "next" {
-                    vsl::accept()
-                } else {
-                    vsl::deny()
-                }
+            && vsl::next().to_string() == "next"
+            && vsl::next().to_debug() == "next" {
+                vsl::accept()
+            } else {
+                vsl::deny()
             }
-        ]
-    }
-)
+        }
+    ]
+}
