+++ src/rules/tests/types/time/main.vsl
@@ -1,17 +1,9 @@
-run_rules!(
-    #{
-        connect: [
-            rule "test_time" || {
-                if custom_time.to_string() != "0" {
-                    return vsl::deny()
-                }
-
-                if custom_time.to_debug() != "0ns" {
-                    return vsl::deny()
-                }
-
-                vsl::accept()
-            }
-        ]
-    }
-)
+#{
+    connect: [
+        rule "test_time" || {
+            print(time.to_string());
+            print(time.to_debug());
+            vsl::accept()
+        }
+    ]
+}
